pyproject.toml
 [build-system]
 requires = ["setuptools>=68", "wheel"]
 build-backend = "setuptools.build_meta"

 [project]
 name = "video-lectures-to-searchable-pdf"
 version = "0.1.0"
 description = "Pipeline to transform webinar videos into searchable slide and transcript PDFs."
 authors = [{ name = "AutoDev Assistant" }]
 readme = "README.md"
 requires-python = ">=3.10"
 dependencies = [
     "fastapi>=0.115.0",
     "uvicorn[standard]>=0.30.0",
    "langchain==0.0.352",
    "langchain-core>=0.1.21,<0.2.0",
    "langsmith<0.1",
    "pydantic>=2.6",
     "pydantic-settings>=2.4",
     "typer>=0.12.0",
     "rich>=13.7",
     "yt-dlp>=2024.4.9",
     "gdown>=5.1.0",
     "ffmpeg-python>=0.2.0",
     "opencv-python>=4.9.0.80",
    "numpy>=1.26,<2",
     "pillow>=10.2.0",
     "pydub>=0.25.1",
     "soundfile>=0.12.1",
     "torch>=2.2.0",
     "transformers>=4.40.0",
     "accelerate>=0.28.0",
     "faster-whisper>=0.10.0",
     "paddlepaddle>=2.6.0",
     "paddleocr>=2.7.0.3",
     "reportlab>=4.1.0",
     "pypdf>=4.0.0",
     "python-multipart>=0.0.9",
     "jinja2>=3.1.0",
 ]

 [project.optional-dependencies]
 dev = ["pytest>=8.2", "pytest-asyncio>=0.23", "ruff>=0.5.5", "mypy>=1.10"]

 [project.scripts]
 vlsp = "app.cli:app"

